[race hub] integrate official webview; add switch; needs improvements
