+++ lib/Screens/session_screen.dart
@@ -17,15 +17,15 @@
  * Copyright (c) 2022-2024, BrightDV
  */
 
+import 'package:boxbox/Screens/SessionWebView/unofficial_webview.dart';
+import 'package:boxbox/Screens/SessionWebView/webview_manager.dart';
 import 'package:boxbox/Screens/free_practice_screen.dart';
 import 'package:boxbox/Screens/race_details.dart';
 import 'package:boxbox/api/event_tracker.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'package:flutter_timer_countdown/flutter_timer_countdown.dart';
+import 'package:hive_flutter/hive_flutter.dart';
 
 class SessionScreen extends StatefulWidget {
   final String sessionFullName;
@@ -38,47 +38,11 @@ class SessionScreen extends StatefulWidget {
 }
 
 class _SessionScreenState extends State<SessionScreen> {
-  final List<ContentBlocker> contentBlockers = [];
-  final CookieManager cookieManager = CookieManager.instance();
-
-  @override
-  void initState() {
-    super.initState();
-    if (!kIsWeb) {
-      ContentBlocker(
-        trigger: ContentBlockerTrigger(
-          urlFilter: ".*",
-          unlessDomain: ["live.planetf1.com"],
-          resourceType: [
-            ContentBlockerTriggerResourceType.SCRIPT,
-            ContentBlockerTriggerResourceType.RAW,
-          ],
-        ),
-        action: ContentBlockerAction(
-          type: ContentBlockerActionType.BLOCK,
-        ),
-      );
-      List<String> selectors = [
-        ".bs-sticky",
-        ".bs-block",
-        ".unic",
-      ];
-      contentBlockers.add(
-        ContentBlocker(
-          trigger: ContentBlockerTrigger(
-            urlFilter: ".*",
-          ),
-          action: ContentBlockerAction(
-            type: ContentBlockerActionType.CSS_DISPLAY_NONE,
-            selector: selectors.join(', '),
-          ),
-        ),
-      );
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
+    bool useOfficialWebview = Hive.box('settings')
+        .get('useOfficialWebview', defaultValue: true) as bool;
+
     int timeBetween(DateTime from, DateTime to) {
       return to.difference(from).inSeconds;
     }
@@ -158,36 +122,9 @@ class _SessionScreenState extends State<SessionScreen> {
               )
             : widget.session.startTime.isBefore(DateTime.now()) &&
                     widget.session.endTime.isAfter(DateTime.now())
-                ? Scaffold(
-                    appBar: AppBar(
-                      backgroundColor: Theme.of(context).colorScheme.onPrimary,
-                      title: Text(
-                        widget.sessionFullName,
-                      ),
-                    ),
-                    body: InAppWebView(
-                      initialUrlRequest: URLRequest(
-                        url: WebUri(
-                          "https://live.planetf1.com/",
-                        ),
-                        headers: {
-                          'User-Agent':
-                              'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:123.0) Gecko/20100101 Firefox/123.0',
-                        },
-                      ),
-                      gestureRecognizers: {
-                        Factory<VerticalDragGestureRecognizer>(
-                          () => VerticalDragGestureRecognizer(),
-                        ),
-                        Factory<HorizontalDragGestureRecognizer>(
-                          () => HorizontalDragGestureRecognizer(),
-                        ),
-                        Factory<ScaleGestureRecognizer>(
-                          () => ScaleGestureRecognizer(),
-                        ),
-                      },
-                    ),
-                  )
+                ? useOfficialWebview
+                    ? WebViewManagerScreen(widget.sessionFullName)
+                    : UnofficialWebviewScreen()
                 : FreePracticeScreen(
                     widget.sessionFullName,
                     int.parse(
@@ -208,8 +145,8 @@ class _SessionScreenState extends State<SessionScreen> {
                 widget.sessionFullName,
               ),
             ),
-            body: widget.session.state == 'upcoming' ||
-                    widget.session.startTime.isAfter(DateTime.now())
+            body: widget.session.state == 'upcoming' && false ||
+                    widget.session.startTime.isAfter(DateTime.now()) && false
                 ? Column(
                     mainAxisAlignment: MainAxisAlignment.center,
                     crossAxisAlignment: CrossAxisAlignment.center,
@@ -281,28 +218,9 @@ class _SessionScreenState extends State<SessionScreen> {
                                   : 'qualifying',
                             ),
                           )
-                    : InAppWebView(
-                        initialUrlRequest: URLRequest(
-                          url: WebUri(
-                            "https://live.planetf1.com/",
-                          ),
-                          headers: {
-                            'User-Agent':
-                                'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:123.0) Gecko/20100101 Firefox/123.0',
-                          },
-                        ),
-                        gestureRecognizers: {
-                          Factory<VerticalDragGestureRecognizer>(
-                            () => VerticalDragGestureRecognizer(),
-                          ),
-                          Factory<HorizontalDragGestureRecognizer>(
-                            () => HorizontalDragGestureRecognizer(),
-                          ),
-                          Factory<ScaleGestureRecognizer>(
-                            () => ScaleGestureRecognizer(),
-                          ),
-                        },
-                      ),
+                    : useOfficialWebview
+                        ? WebViewManagerScreen(widget.sessionFullName)
+                        : UnofficialWebviewScreen(),
           );
   }
 }
